+++ FirebaseAuth/Sources/Swift/Utilities/AuthRecaptchaVerifier.swift
@@ -125,36 +125,85 @@
         // No recaptcha on internal build system.
         return actionString
       #else
-        return try await withCheckedThrowingContinuation { continuation in
-          FIRRecaptchaGetToken(siteKey, actionString,
-                               "NO_RECAPTCHA") { (token: String, error: Error?,
-                                                  linked: Bool, actionCreated: Bool) in
-              guard linked else {
-                continuation.resume(throwing: AuthErrorUtils.recaptchaSDKNotLinkedError())
-                return
-              }
-              guard actionCreated else {
-                continuation.resume(throwing: AuthErrorUtils.recaptchaActionCreationFailed())
-                return
-              }
-              if let error {
-                continuation.resume(throwing: error)
-                return
-              } else {
-                if token == "NO_RECAPTCHA" {
-                  AuthLog.logInfo(code: "I-AUT000031",
-                                  message: "reCAPTCHA token retrieval failed. NO_RECAPTCHA sent as the fake code.")
-                } else {
-                  AuthLog.logInfo(
-                    code: "I-AUT000030",
-                    message: "reCAPTCHA token retrieval succeeded."
-                  )
-                }
-                continuation.resume(returning: token)
-              }
-          }
+
+        let (token, error, linked, actionCreated) = await recaptchaToken(
+          siteKey: siteKey,
+          actionString: actionString,
+          fakeToken: "NO_RECAPTCHA"
+        )
+
+        guard linked else {
+          throw AuthErrorUtils.recaptchaSDKNotLinkedError()
         }
+        guard actionCreated else {
+          throw AuthErrorUtils.recaptchaActionCreationFailed()
+        }
+        if let error {
+          throw error
+        }
+        if token == "NO_RECAPTCHA" {
+          AuthLog.logInfo(code: "I-AUT000031",
+                          message: "reCAPTCHA token retrieval failed. NO_RECAPTCHA sent as the fake code.")
+        } else {
+          AuthLog.logInfo(
+            code: "I-AUT000030",
+            message: "reCAPTCHA token retrieval succeeded."
+          )
+        }
+        return token
       #endif // !(COCOAPODS || SWIFT_PACKAGE)
+    }
+
+    private static var recaptchaClient: (any RCARecaptchaClientProtocol)?
+
+    #if COCOAPODS || SWIFT_PACKAGE // No recaptcha on internal build system.
+      private func recaptchaToken(siteKey: String,
+                                  actionString: String,
+                                  fakeToken: String) async -> (token: String, error: Error?,
+                                                               linked: Bool, actionCreated: Bool) {
+        if let recaptchaClient {
+          return await retrieveToken(
+            actionString: actionString,
+            fakeToken: fakeToken,
+            recaptchaClient: recaptchaClient
+          )
+        }
+
+        if let recaptcha =
+          NSClassFromString("RecaptchaEnterprise.RCARecaptcha") as? RCARecaptchaProtocol.Type {
+          do {
+            let client = try await recaptcha.fetchClient(withSiteKey: siteKey)
+            recaptchaClient = client
+            return await retrieveToken(
+              actionString: actionString,
+              fakeToken: fakeToken,
+              recaptchaClient: client
+            )
+          } catch {
+            return ("", error, true, true)
+          }
+        } else {
+          // RecaptchaEnterprise not linked.
+          return ("", nil, false, false)
+        }
+      }
+    #endif // (COCOAPODS || SWIFT_PACKAGE)
+
+    private func retrieveToken(actionString: String,
+                               fakeToken: String,
+                               recaptchaClient: RCARecaptchaClientProtocol) async -> (token: String,
+                                                                                      error: Error?,
+                                                                                      linked: Bool,
+                                                                                      actionCreated: Bool) {
+      if let recaptchaAction =
+        NSClassFromString("RecaptchaEnterprise.RCAAction") as? RCAActionProtocol.Type {
+        let action = recaptchaAction.init(customAction: actionString)
+        let token = try? await recaptchaClient.execute(withAction: action)
+        return (token ?? "NO_RECAPTCHA", nil, true, true)
+      } else {
+        // RecaptchaEnterprise not linked.
+        return ("", nil, false, false)
+      }
     }
 
     func retrieveRecaptchaConfig(forceRefresh: Bool) async throws {