+++ FirebaseAuth/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 11.9.0
+- [changed] Using reCAPTCHA Enterprise and Firebase Auth requires reCAPTCHA
+  Enterprise 18.7.0 or later.
+
+# 11.8.1
+- [fixed] Suppress deprecation build warning introduced in 11.8.0.
+
+# 11.8.0
+- [added] Added `ActionCodeSettings.linkDomain` to customize the Firebase Hosting link domain
+  that is used in out-of-band email action flows.
+- [deprecated] Deprecated `ActionCodeSettings.dynamicLinkDomain`.
+
 # 11.7.0
 - [fixed] Fix Multi-factor session crash on second Firebase app. (#14238)
 - [fixed] Updated most decoders to be consistent with Firebase 10's behavior