+++ FirebaseMessaging/CHANGELOG.md
@@ -1,4 +1,7 @@
-# Unreleased
+# 11.9.0
+- [fixed] Migrate FCM codebase to new NSKeyedUnarchiver APIs. (#14424).
+
+# 11.8.0
 - [fixed] Don't cache FCM registration token operations. (#14352).
 
 # 11.5.0